% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropyears.R
\docType{methods}
\name{cropyears}
\alias{cropyears}
\alias{cropyears,Date-method}
\alias{cropyears,RasterBrick-method}
\alias{cropyears,RasterList-method}
\alias{cropyears,RasterStack-method}
\alias{cropyears,character-method}
\alias{cropyears,data.frame-method}
\alias{cropyears,list-method}
\alias{cropyears,numeric-method}
\title{Crop a monthly or daily or sub-daily time series to a series starting from the beginning of the year and ending of the end of an year.}
\usage{
cropyears(object = NULL, start, end, months_offset = c(0, 6), ...)

cropyears(object = NULL, start, end, months_offset = c(0, 6), ...)

\S4method{cropyears}{Date}(object = NULL, start, end, months_offset = c(0,
  6), ...)

\S4method{cropyears}{data.frame}(object = NULL, timestamp = "Date", ...)

\S4method{cropyears}{character}(object = NULL, format_date = "X\%Y.\%m.\%d",
  ...)

\S4method{cropyears}{numeric}(object = NULL, format_date = "X\%Y.\%m.\%d",
  ...)

\S4method{cropyears}{RasterStack}(object = NULL,
  format_date = "X\%Y.\%m.\%d", ...)

\S4method{cropyears}{RasterBrick}(object = NULL,
  format_date = "X\%Y.\%m.\%d", ...)

\S4method{cropyears}{list}(object = NULL, start, end, months_offset = c(0,
  6), ...)

\S4method{cropyears}{RasterList}(object = NULL,
  format_date = "X\%Y.\%m.\%d", ...)
}
\arguments{
\item{object}{time series to crop (or object containing time series to crop)}

\item{start}{start date of time series, used  if \code{object} is missing or \code{NULL}}

\item{end}{end  date of time series, used  if \code{object} is missing or \code{NULL}}

\item{months_offset}{offset expressed in number of months from January 1st, used if the year is chosen to start from a different from January.}

\item{...}{further arguments}

\item{timestamp}{character string containing the data frame field name for the date. Default is \code{"Date"}.}

\item{format_date}{string character used for date. Default is \code{"X\%Y.\%m.\%d"}.}
}
\description{
It crops a time series to a one staring from the beginning of the year and ending at the end of the year.
}
\examples{

start <- as.Date("1988-09-09")
end <-   as.Date("2016-02-02")
cropdates <- cropyears(start=start,end=end)

start <- as.Date("1979-01-01")
end <-   as.Date("2016-12-31")
cropdates <- cropyears(start=start,end=end)


format_date <- "X\%Y.\%m.\%d"
dates <- start+1:700
vecn <- rnorm(700)
names(vecn) <- as.character(dates,format=format_date)
cropvect <- cropyears(vecn,format_date=format_date)


t2m <- stack(system.file("map/t2m_daily.grd",package="rasterList"))
cropt2m <- cropyears(t2m)




}

