% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.linktolist.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,LinkToList,ANY,ANY-method}
\alias{[[}
\alias{[[,LinkToList,ANY,missing-method}
\alias{length}
\alias{length,LinkToList-method}
\title{Extract parts of \code{\link{LinkToList-class}}}
\usage{
\S4method{[}{LinkToList,ANY,ANY}(x, i, j, ..., forceInMemory = FALSE,
  filename = rasterTmpFile("ll"), nperfile = ceiling(length(i)/10),
  drop = TRUE)

\S4method{[[}{LinkToList,ANY,missing}(x, i, j, ...)

\S4method{length}{LinkToList}(x)
}
\arguments{
\item{x}{a \code{\link{LinkToList-class}} object}

\item{i}{indices of the element to get}

\item{j, drop, ...}{further arguments}

\item{forceInMemory}{logical value. Deafault is \code{FALSE}. If it is \code{TRUE}, method return a \code{list} saved in memomey, otheriwise it can return a  \code{\link{LinkToList-class}} object}

\item{filename, nperfile}{arguments passed to \link{linkToList} if \code{forceInMemory==FALSE}.}
}
\description{
Extract the ith element of a  \code{\link{LinkToList-class}}  object.
}
\examples{

set.seed(123)

x <- list()
n <- 1000 
for (i in 1:n) {

		x[[i]] <- rnorm(100)
}

ll <- linkToList(x)

filename <- rasterTmpFile("test100")

ll2 <- linkToList(x,nperfile=100)

ll2[c(9,10,800)]
x[c(9,10,800)]

x[[28]]
ll2[[28]]

}
\seealso{
\code{\link{linkToList}}
}
