% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.RasterList.R
\name{is.RasterList}
\alias{is.RasterList}
\title{Is a \code{\link{RasterList-class}} object ?}
\usage{
is.RasterList(x)
}
\arguments{
\item{x}{a valid object}
}
\description{
Is a \code{\link{RasterList-class}} object ?
}
\examples{

r <- rasterList()
is.RasterList(r)
rr <- raster()
is.RasterList(rr)

f <- system.file("external/test.grd", package="raster")
ra <- rasterList(f)
is.RasterList(rr)



}

