% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.RasterListAble.R
\name{is.RasterListAble}
\alias{is.RasterListAble}
\title{Can it be coerced to a \code{\link{RasterList-class}} object ?}
\usage{
is.RasterListAble(x)
}
\arguments{
\item{x}{a valid object}
}
\description{
Can it be coerced to a \code{\link{RasterList-class}} object ?
}
\examples{

r <- rasterList()
is.RasterListAble(r)
rr <- raster()
is.RasterListAble(rr)

f <- system.file("external/test.grd", package="raster")
ra <- stack(c(f,f))
is.RasterListAble(ra)



}
