% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkToList.R
\docType{methods}
\name{linkToList}
\alias{as.LinkToList}
\alias{linkToList}
\alias{linkToList,LinkToList-method}
\alias{linkToList,list-method}
\title{Creates a \code{\link{LinkToList-class}}  object}
\usage{
linkToList(object = NULL, filename = "default", FUN = NULL,
  nperfile = 100, length = 200, ...)

as.LinkToList(object, ...)

linkToList(object = NULL, filename = "default", FUN = NULL,
  nperfile = 100, length = 200, ...)

\S4method{linkToList}{LinkToList}(object, filename = "default", FUN = NULL,
  nperfile = "not.used", suffix = "default", ...)

\S4method{linkToList}{list}(object, filename = "default", FUN = NULL,
  nperfile = length, length = length(object), single.element = TRUE, ...)
}
\arguments{
\item{object}{the object to coerce}

\item{filename}{a character name using as a root for the file names.}

\item{FUN}{function that can be optionally processed per each element of the \code{\link{LinkToList-class}} object}

\item{nperfile}{maximum number of \code{*.rda} files}

\item{length}{length of the \code{\link{LinkToList-class}}  object  to be created (e.g. if \code{object} is \code{NULL} or \code{list()})}

\item{...}{further arguments , e.g. for \code{\link{save}}}

\item{suffix}{suffix that can be optionally applied if \code{FUN} is not \code{NULL}.}

\item{single.element}{logical parameter. Dafault is \code{TRUE}. If it is \code{TRUE} , \code{FUN} is appled to ach single element of the list, otharwise \code{FUN} is applied to teach entire \code{\link{RasterList-class}} block of elements.}
}
\description{
The method \code{LinkToList} is the constructor of a  \code{\link{LinkToList-class}} from a generic \code{object}.
}
\details{
This methods is used to write a list (e.g. the one of a \code{\link{RasterList-class}} object) into a set of files in order to save memory, 
the \code{\link{LinkToList-class}} object.
}
\examples{



set.seed(123)

x <- list()
n <- 1000 
for (i in 1:n) {

		x[[i]] <- rnorm(100)
}

ll <- linkToList(x)

filename <- rasterTmpFile("test100")

ll2 <- linkToList(x,nperfile=100)

is.list(ll2)

ll3 <- linkToList(ll2,nperfile=100,FUN=mean)
ll3a <- linkToList(ll2,nperfile=100,FUN=mean,filename="/tmp/mapname.asc")
ll4 <- linkToList(x,nperfile=100,FUN=mean)

}

