% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA_01_classes.R
\docType{class}
\name{list-class}
\alias{list-class}
\alias{list.LinkToList}
\title{S4 declaration for \code{list} S4 class object}
\description{
S4 declaration for \code{list} S4 class object
}
\examples{

showClass("list")

}
\seealso{
\code{\link{as.list}},\code{\link{list}}
}
\keyword{classes}
