% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapplyLinkToList.R
\name{mapplyLinkToList}
\alias{mapplyLinkToList}
\title{Use of \link{mapply} functions with \link{LinkToList-class} objects}
\usage{
mapplyLinkToList(..., filename = "default")
}
\arguments{
\item{...}{arguments}

\item{filename}{argument passed to \code{\link{linkToList}}.}
}
\description{
Use of \link{mapply} functions with \link{LinkToList-class} objects
}
\examples{

set.seed(123)

x <- list()
n <- 1000 
for (i in 1:n) {

		x[[i]] <- rnorm(100)
}

ll <- linkToList(x)

filename <- rasterTmpFile("test100")

ll2 <- linkToList(x,nperfile=100)

is.list(ll2)

ll3 <- linkToList(ll2,nperfile=100,FUN=mean)
ll3a <- linkToList(ll2,nperfile=100,FUN=mean,filename="/tmp/mapname.asc")
ll4 <- linkToList(x,nperfile=100,FUN=mean)


ll5 <- mapplyLinkToList(ll3,ll4,3,FUN=sum)

ll5a <- mapplyLinkToList(as.list(ll3),as.list(ll4),3,FUN=sum)

as.list(ll5)
ll5a

}

