% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncell.R
\docType{methods}
\name{ncell,RasterList-method}
\alias{ncell,RasterList-method}
\alias{ncell}
\title{Number of cells in a \code{\link{RasterList-class}} object}
\usage{
\S4method{ncell}{RasterList}(x)
}
\arguments{
\item{x}{a  \code{\link{RasterList-class}} object}
}
\description{
Number of cells in a \code{\link{RasterList-class}} object
}
\details{
This function is an implementation of the \code{ncell} created in the \code{raster} package. The number of cells corresponds to the number of elements of the associeted list.
}
\examples{

f <- system.file("external/test.grd", package="raster")
ra <- rasterList(f)
is.RasterList(ra)
ncell(ra)


}
