% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAClasses.R
\docType{class}
\name{RasterList-class}
\alias{RasterList-class}
\title{RasterList-class}
\description{
Class \code{RasterList}
}
\details{
It contains \code{\link{RasterLayer-class}} with the following adjoint slots:
 \describe{
    \item{\code{list}:}{a list of generic object whese length corresponds to the number of cells. Each \code{list} element for each cell; }

\item{\code{name}:}{an identification name of the object. Default is \code{NA}.}
 }

This class inherits the \code{\link{RasterLayer-class}} class considering each pixel of the raster is a generic object.
}
\examples{

showClass("RasterList")

}
\author{
Emanuele Cordano
}
\seealso{
\code{\link{raster}},\code{\link{Raster-class}}
}
\keyword{classes}

