% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.linktolist.R
\docType{methods}
\name{[<-}
\alias{[<-}
\alias{[<-,LinkToList,ANY,ANY,ANY-method}
\alias{[[<-}
\alias{[[<-,LinkToList,ANY,missing,ANY-method}
\title{Replace parts of \code{\link{LinkToList-class}}}
\usage{
\S4method{[}{LinkToList,ANY,ANY,ANY}(x, i, j, ...) <- value

\S4method{[[}{LinkToList,ANY,missing,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{a \code{\link{LinkToList-class}} object}

\item{i}{indices of the element to get}

\item{j, ...}{further arguments}

\item{value}{a generic \code{list} or a \code{\link{LinkToList-class}} obbject containing the elements to be replaced,. It should of length 1 or equal to \code{i}.}
}
\description{
Extract the ith element of a  \code{\link{LinkToList-class}}  object.
}
\examples{

set.seed(123)

x <- list()
n <- 1000 
for (i in 1:n) {

		x[[i]] <- rnorm(100)
}

ll <- linkToList(x)

filename <- rasterTmpFile("test100")

ll2 <- linkToList(x,nperfile=100)

ll2[c(9,10,800)]
ll2[c(9,800,100)]
x[c(9,10,800)] <- list("a","b","c")
ll2[c(9,10,800)] <- list("a","b","c")
if (identical(x,as.list(ll2))==FALSE) stop("x and ll2 should be identical!! Something went wrong!!")
  
}
\seealso{
\code{\link{rasterList}}
}

