% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\docType{methods}
\name{stack,RasterList-method}
\alias{stack}
\alias{stack,RasterList-method}
\title{Creates a \code{\link{RasterStack-class}}  object from a \code{\link{RasterList-class}}}
\usage{
\S4method{stack}{RasterList}(x, ...)
}
\arguments{
\item{x}{a \code{rasterList-class} object}

\item{...}{further arguments for \code{\link{rasterList}}}
}
\description{
The method transforms a \code{\link{RasterList-class}} into a \code{\link{RasterStack-class}} in case of the list elements are numeric vectors.
}
\examples{

f <- system.file("external/test.grd", package="raster")

## Creates a simple generic RasterList

rl <- rasterList(f) 

list <- as.list(as.vector(rl))
list <- lapply(X=list,FUN=function (x) {c(x,x+10,x+15)}) 

rl <- rasterList(rl,list=list,object.name="test")

ss <- stack(rl)

il <- 8331
list[[il]] <- numeric(0)
rla <- rasterList(rl,list=list,object.name="test2")
sa <- stack(rla) 
}
\seealso{
\code{\link{rasterList}}
}

