% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.list.R
\docType{methods}
\name{as.list}
\alias{as.list}
\alias{as.list,LinkToList-method}
\alias{LinkToList2list}
\title{Coercion of a \link{LinkToList-class} object to a generic \link{list} object.}
\usage{
\S4method{as.list}{LinkToList}(x, ...)

LinkToList2list(x, ...)
}
\arguments{
\item{x}{a \link{LinkToList-class} object}

\item{...}{forther arguments for \code{\link{[}} method}
}
\description{
Coercion of a \link{LinkToList-class} object to a generic \link{list} object.
}
