% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA_02_classes.R
\docType{class}
\name{LinkToList-class}
\alias{LinkToList-class}
\alias{LinkToList}
\title{S4 declaration for \code{LinkToList-class} S4 class object}
\description{
S4 declaration for \code{LinkToList-class} S4 class object
}
\examples{

showClass("LinkToList")

}
\seealso{
\code{\link{LinkToList}}
}
\keyword{classes}
