% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_LinkToList.R
\docType{methods}
\name{show,LinkToList-method}
\alias{show,LinkToList-method}
\alias{show,LinkToList}
\title{Print or Show S4 method for a \code{\link{LinkToList-class}} object}
\usage{
\S4method{show}{LinkToList}(object)
}
\arguments{
\item{object}{a \code{\link{LinkToList-class}} object}
}
\description{
Prints a \code{\link{LinkToList-class}} object
}
